+++ CHANGELOG.md
@@ -20,7 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [#1460] When importing parameter assets in MA Parameters, "local only" parameters were incorrectly treated as
   "animator only"
 - [#1489] Fixed compatibility issues between `Merge Blend Tree` or reactive components and MMD worlds.
-  See [documentation](https://modular-avatar.nadena.dev/docs/general-behavior/mmd) for details on the new handling. 
+  - See [documentation](https://modular-avatar.nadena.dev/docs/general-behavior/mmd) for details on the new handling.
+  - To disable this behavior, attach the new `MA VRChat Settings` component to any object on your avatar and disable the appropriate setting.
 - [#1501] Unity keyboard shortcuts don't work when editing text fields on the MA Parameters component 
 - [#1410] Motion overrides on synced layers are not updated for Bone Proxy/Merge Armature object movement
 - [#1504] The internal `DelayDisable` layer no longer references unnecessary objects in some situations